clean the file before loading into pandas
