+++ pyproject.toml
@@ -20,6 +20,7 @@ psycopg2 = "^2.9.6"
 pandas-stubs = "^2.1.1.230928"
 types-python-dateutil = "^2.8.19.14"
 types-openpyxl = "^3.1.0.26"
+tqdm = "^4.66.4"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.1.0"
